Move rotate, flip and delete to own class
These are the clean and simple methods that are not related to the
manipulate toolbar. It therefore makes sense to outsource them to a
different class.

This is also a first step in re-implementing this in a cleaner more
signal-based way.
