+++ tests/manipulate_test.py
@@ -3,11 +3,7 @@
 
 import os
 import shutil
-import tempfile
-from time import sleep
 from unittest import main
-
-from PIL import Image
 
 from vimiv_testcase import VimivTestCase, compare_images
 
@@ -22,149 +18,6 @@ class ManipulateTest(VimivTestCase):
         shutil.copytree("vimiv/testimages", "vimiv/testimages_man")
         cls.init_test(cls, ["vimiv/testimages_man/arch-logo.png"])
         cls.manipulate = cls.vimiv["manipulate"]
-
-    def test_get_manipulated_images(self):
-        """Get images to manipulate."""
-        # Nothing marked -> current image
-        expected_images = [self.vimiv.paths[self.vimiv.index]]
-        received_images = self.manipulate.get_manipulated_images("test")
-        self.assertEqual(expected_images, received_images)
-        # Marked images
-        marked = ["arch-logo.png", "symlink_to_image"]
-        marked = [os.path.abspath(image) for image in marked]
-        self.vimiv["mark"].marked = marked
-        received_images = self.manipulate.get_manipulated_images("test")
-        self.assertEqual(marked, received_images)
-        # Reset mark
-        self.vimiv["mark"].marked = []
-
-    def test_delete_undelete(self):
-        """Delete and undelete images from within manipulate."""
-        # Delete
-        self.assertTrue(os.path.exists("arch-logo.png"))
-        self.manipulate.delete()
-        self.assertFalse(os.path.exists("arch-logo.png"))
-        self.assertEqual(self.vimiv.paths[self.vimiv.index],
-                         os.path.abspath("arch_001.jpg"))
-        # Undelete
-        self.manipulate.undelete("arch-logo.png")
-        self.assertTrue(os.path.exists("arch-logo.png"))
-        self.assertIn(os.path.abspath("arch-logo.png"), self.vimiv.paths)
-
-    def test_fail_delete_undelete(self):
-        """Fail deleting and undeleting images."""
-        # Restore a non-existent file
-        self.manipulate.undelete("foo_bar_baz")
-        self.check_statusbar(
-            "ERROR: Could not restore foo_bar_baz, file does not exist")
-        # Restore a directory
-        tmpdir = tempfile.mkdtemp()
-        basename = os.path.basename(tmpdir)
-        self.vimiv["manipulate"].trash_manager.delete(tmpdir)
-        self.manipulate.undelete(basename)
-        self.check_statusbar(
-            "ERROR: Could not restore %s, directories are not supported"
-            % (basename))
-        # Restore a file without a top directory
-        tmpdir = tempfile.mkdtemp()
-        extra_directory = os.path.join(tmpdir, "extra_directory")
-        new_file = os.path.join(extra_directory, "new_file")
-        os.mkdir(extra_directory)
-        with open(new_file, "w") as f:
-            f.write("")
-        self.vimiv["manipulate"].trash_manager.delete(new_file)
-        basename = os.path.basename(new_file)
-        shutil.rmtree(tmpdir)
-        self.manipulate.undelete(basename)
-        self.check_statusbar(
-            "ERROR: Could not restore %s, directory is not accessible"
-            % (basename))
-
-    def test_rotate(self):
-        """Rotate image."""
-        # Before
-        pixbuf = self.vimiv["image"].get_pixbuf()
-        is_portrait = pixbuf.get_width() < pixbuf.get_height()
-        self.assertFalse(is_portrait)
-        # Rotate
-        self.manipulate.rotate(1, True)
-        pixbuf = self.vimiv["image"].get_pixbuf()
-        is_portrait = pixbuf.get_width() < pixbuf.get_height()
-        self.assertTrue(is_portrait)
-        # Rotate the file
-        with Image.open(self.vimiv.paths[self.vimiv.index]) as im:
-            self.assertFalse(im.width < im.height)
-        self.manipulate.run_simple_manipulations()
-        while self.manipulate.threads_running():
-            sleep(0.05)
-        with Image.open(self.vimiv.paths[self.vimiv.index]) as im:
-            self.assertTrue(im.width < im.height)
-        # Rotate back
-        self.manipulate.rotate(1, True)
-        pixbuf = self.vimiv["image"].get_pixbuf()
-        is_portrait = pixbuf.get_width() < pixbuf.get_height()
-        self.assertFalse(is_portrait)
-        # Rotate the file
-        self.manipulate.run_simple_manipulations()
-        while self.manipulate.threads_running():
-            sleep(0.05)
-        with Image.open(self.vimiv.paths[self.vimiv.index]) as im:
-            self.assertFalse(im.width < im.height)
-        # Fail because of no paths
-        backup = list(self.vimiv.paths)
-        self.vimiv.paths = []
-        self.manipulate.rotate(1, False)
-        self.check_statusbar("ERROR: No image to rotate")
-        self.vimiv.paths = backup
-        ##################
-        #  Command line  #
-        ##################
-        # Rotate
-        self.run_command("rotate 1")
-        pixbuf = self.vimiv["image"].get_pixbuf()
-        is_portrait = pixbuf.get_width() < pixbuf.get_height()
-        self.assertTrue(is_portrait)
-        # Rotate back
-        self.run_command("rotate -1")
-        pixbuf = self.vimiv["image"].get_pixbuf()
-        is_portrait = pixbuf.get_width() < pixbuf.get_height()
-        self.assertFalse(is_portrait)
-        # Fail because of invalid argument
-        self.run_command("rotate value")
-        self.check_statusbar(
-            "ERROR: Argument for rotate must be of type integer")
-
-    def test_flip(self):
-        """Flip image."""
-        # Flip
-        pixbuf_before = self.vimiv["image"].get_pixbuf()
-        self.manipulate.flip(1, False)
-        pixbuf_after = self.vimiv["image"].get_pixbuf()
-        self.assertNotEqual(pixbuf_before, pixbuf_after)
-        # Flip back
-        self.manipulate.flip(1, False)
-        pixbuf_after_2 = self.vimiv["image"].get_pixbuf()
-        self.assertNotEqual(pixbuf_after, pixbuf_after_2)
-        # Fail because of no paths
-        backup = list(self.vimiv.paths)
-        self.vimiv.paths = []
-        self.manipulate.flip(1, False)
-        self.check_statusbar("ERROR: No image to flip")
-        self.vimiv.paths = backup
-        ##################
-        #  Command line  #
-        ##################
-        # Flip
-        self.run_command("flip 1")
-        pixbuf_after_3 = self.vimiv["image"].get_pixbuf()
-        self.assertNotEqual(pixbuf_after_2, pixbuf_after_3)
-        # Flip back
-        self.run_command("flip 1")
-        pixbuf_after_4 = self.vimiv["image"].get_pixbuf()
-        self.assertNotEqual(pixbuf_after_3, pixbuf_after_4)
-        # Fail because of invalid argument
-        self.run_command("flip value")
-        self.check_statusbar("ERROR: Argument for flip must be of type integer")
 
     def test_toggle(self):
         """Toggle manipulate."""
